feat: move `AllowedFieldTypes` and `AllowedQuestionTypes` to `typing.py`
