+++ tests/client/conftest.py
@@ -21,7 +21,7 @@ from argilla.client.sdk.datasets.models import TaskType
 from datasets import Dataset
 
 if TYPE_CHECKING:
-    from argilla.client.feedback.schemas import AllowedFieldTypes, AllowedQuestionTypes
+    from argilla.client.feedback.typing import AllowedFieldTypes, AllowedQuestionTypes
 
 from argilla.client.feedback.schemas import (
     FeedbackRecord,
