+++ tests/client/feedback/training/test_trainer.py
@@ -17,7 +17,7 @@ from typing import TYPE_CHECKING, List, Union
 import pytest
 
 if TYPE_CHECKING:
-    from argilla.client.feedback.schemas import AllowedFieldTypes, AllowedQuestionTypes
+    from argilla.client.feedback.typing import AllowedFieldTypes, AllowedQuestionTypes
 
 import shutil
 import sys
